Complete behavioral test-suite refactor
