+++ tests/test_canonical_problems_notebook.py
@@ -14,12 +14,6 @@ def notebook() -> dict:
     return json.loads(NOTEBOOK_PATH.read_text())
 
 
-def notebook_source() -> str:
-    return "\n".join(
-        "".join(cell.get("source", [])) for cell in notebook()["cells"]
-    )
-
-
 def notebook_cell(data: dict, cell_id: str) -> dict:
     matches = [cell for cell in data["cells"] if cell.get("id") == cell_id]
     if len(matches) != 1:
@@ -39,74 +33,7 @@ def binary_states(size: int) -> list[tuple[int, ...]]:
     return list(itertools.product((0, 1), repeat=size))
 
 
-def assert_source_contains(
-    test_case: unittest.TestCase, source: str, marker: str
-) -> None:
-    test_case.assertTrue(marker in source, f"Missing notebook source marker: {marker!r}")
-
-
-def assert_source_excludes(
-    test_case: unittest.TestCase, source: str, marker: str
-) -> None:
-    test_case.assertFalse(marker in source, f"Forbidden notebook source marker: {marker!r}")
-
-
-class CanonicalProblemsNotebookSourceTests(unittest.TestCase):
-    def test_notebook_has_tutorial_structure_and_clean_room_references(self) -> None:
-        source = notebook_source()
-        required_sections = (
-            "## Setup",
-            "## Learning objectives",
-            "## Prerequisites",
-            "## Number partitioning",
-            "## Max-Cut",
-            "## Minimum vertex cover",
-            "## Practice checkpoints",
-            "## Summary",
-            "## References",
-        )
-
-        for section in required_sections:
-            with self.subTest(section=section):
-                assert_source_contains(self, source, section)
-
-        assert_source_contains(self, source, "https://doi.org/10.1287/educ.2025.0288")
-        assert_source_contains(
-            self,
-            source,
-            "https://github.com/arulrhikm/Solving-QUBOs-on-Quantum-Computers",
-        )
-        assert_source_contains(self, source.lower(), "original julia derivations")
-
-    def test_notebook_uses_public_deterministic_credential_free_path(self) -> None:
-        source = notebook_source()
-
-        required_markers = (
-            "QUBO.ExactSampler.Optimizer",
-            "evaluate_jump_objective",
-            "assert_all_states_match",
-            "exact_sampler_optima!",
-            "fixed_positions",
-            "partition_weights = [1, 3, 4, 8]",
-            "(3, 4, 3),",
-            "cover_edges = [(1, 2), (1, 3), (2, 3), (3, 4)]",
-            "@assert partition_best_energy == 0",
-            "@assert maxcut_best_energy == -7",
-            "@assert cover_best_energy == 2",
-            "@assert length(partition_optima) == 2",
-            "@assert length(maxcut_optima) == 4",
-            "@assert length(cover_optima) == 2",
-        )
-        for marker in required_markers:
-            with self.subTest(marker=marker):
-                assert_source_contains(self, source, marker)
-
-        assert_source_excludes(self, source, "unsafe_backend")
-
-        for out_of_scope_name in ("Qiskit", "OpenQAOA", "D-Wave", "token", "API key"):
-            with self.subTest(out_of_scope_name=out_of_scope_name):
-                assert_source_excludes(self, source, out_of_scope_name)
-
+class CanonicalProblemsNotebookOutputTests(unittest.TestCase):
     def test_notebook_commits_reproducible_teaching_outputs(self) -> None:
         data = notebook()
         code_cells = [cell for cell in data["cells"] if cell["cell_type"] == "code"]
@@ -158,17 +85,6 @@ class CanonicalProblemsNotebookSourceTests(unittest.TestCase):
                     marker in serialized_outputs,
                     f"Forbidden marker in committed outputs: {marker!r}",
                 )
-
-    def test_notebook_and_verifier_are_linked(self) -> None:
-        readme = (REPO_ROOT / "README.md").read_text()
-        makefile = (REPO_ROOT / "Makefile").read_text()
-
-        self.assertIn("notebooks_jl/7-CanonicalProblems.ipynb", readme)
-        self.assertIn("make verify-canonical-problems-julia", readme)
-        self.assertIn("verify-canonical-problems-julia:", makefile)
-        self.assertIn(
-            'NOTEBOOKS="$(CANONICAL_PROBLEMS_JULIA_NOTEBOOK)"', makefile
-        )
 
 
 class CanonicalProblemsMathematicalInvariantTests(unittest.TestCase):
